Extract CoverView from StoryView
