+++ src/components/story-view/story-view.tsx
@@ -1,14 +1,13 @@
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 import classNames from 'classnames';
 
 // FPCC
-import { Bookmark, FVStory } from '../common/data/types';
+import { FVStory } from '../common/data/types';
 import Modal from '../common/modal/modal';
 import FvImage from '../common/image/image';
 import AudioControl from '../common/audio-control/audio-control';
 import { convertJsonToComponent } from '../common/convert-json/convert-json';
-import CopyButton from '../common/copy-button/copy-button';
-import BookmarkButton from '../common/bookmark-button/bookmark-button';
+import CoverView from '../story-view/cover-view';
 
 export interface StoryViewProps {
   story: FVStory;
@@ -20,22 +19,12 @@ export function StoryView({ story, goBack }: Readonly<StoryViewProps>) {
   const [showPictureModal, setShowPictureModal] = useState<boolean>(false);
   const [pictureUrl, setPictureUrl] = useState<string>('');
 
-  const bookmark: Bookmark = useMemo(() => {
-    return {
-      id: story.id,
-      type: 'story',
-      definition: story?.titleTranslation ?? '',
-      name: story.title ?? '',
-      hasAudio: story.relatedAudio?.length !== 0,
-      url: `${window.location.pathname}#${story.id}`,
-      timestamp: new Date(),
-    };
-  }, [story]);
-
   return (
     <>
       <div>
-        {currentPage === -2 && title()}
+        {currentPage === -2 && (
+          <CoverView story={story} startReading={() => setCurrentPage(-1)} />
+        )}
         {currentPage === -1 && intro()}
         {currentPage !== -2 && currentPage !== -1 && page()}
       </div>
@@ -47,43 +36,6 @@ export function StoryView({ story, goBack }: Readonly<StoryViewProps>) {
       )}
     </>
   );
-
-  function title() {
-    return (
-      <div className="flex items-center border border-gray-300 shadow-lg p-5 rounded-lg max-w-2xl mx-auto mb-5">
-        <div className="space-y-5">
-          {story?.relatedImages[0] && (
-            <FvImage
-              className="h-[58vh] w-[90vw] object-contain"
-              src={story?.relatedImages[0].original.path ?? ''}
-              alt={story?.title ?? ''}
-            />
-          )}
-
-          <div className="space-y-1">
-            <div className="text-lg md:text-2xl font-bold">{story?.title}</div>
-            <div className="text-sm md:text-base">
-              {story?.titleTranslation}
-            </div>
-          </div>
-          <div className="block space-y-2">
-            <CopyButton text={story?.title} />
-            {/* hiding share button FW-5780 {shareButton()} */}
-            <BookmarkButton bookmark={bookmark} />
-          </div>
-          <button
-            onClick={() => setCurrentPage(-1)}
-            type="button"
-            className={classNames(
-              'w-full btn-contained bg-color-alphabet-light'
-            )}
-          >
-            Start Reading
-          </button>
-        </div>
-      </div>
-    );
-  }
 
   function intro() {
     return (
